Refactor components and update docs
